refactor(portal): extract tracks section into TracksSection.svelte
Move the "your tracks" section — list rows, the inline metadata edit form,
the audio-replace + version-history flows, and their three tied modals — out
of the portal page into a co-located TracksSection.svelte. Parent shrinks
3785 -> 2126 lines.

The shared tracks/albums caches and their loaders stay in the parent (the data
section's PdsBackfillControl + export gate read `tracks`/`tracksTotal`, and the
albums section + the edit form's AlbumSelect both read `albums`), passed down
as props with onLoadMore / onTracksChanged callbacks so the child never owns or
duplicates the fetch. player/uploader/toast stores are imported directly.
